unify: move more code to dedicated classes
